+++ src/dashtools/dashboard/callbacks/deployPage_callbacks.py
@@ -8,13 +8,14 @@ try:
     import alerts
     import tree
     from dashtools.dashboard.pages import deployPage
+    from dashtools.deploy import fileUtils, gitUtils, herokuUtils
 except ModuleNotFoundError:
     from .. import tree
     from .. import alerts
     from ..pages import deployPage
+    from ...deploy import fileUtils, gitUtils, herokuUtils
 
 from dash_iconify import DashIconify
-from dashtools.deploy import fileUtils, gitUtils, herokuUtils
 
 
 def generate_callbacks(app: Dash):
@@ -184,18 +185,20 @@ def generate_callbacks(app: Dash):
     )
     def file_explorer_callback(n, force_tree_update, filepath: os.PathLike):
         # Initial callbacks
+        trigger_id = ctx.triggered_id
         empty_div = dmc.Center([
             html.H3("Open a File to Continue", style={
                     'opacity': '10%', 'padding-top': '50px'})
         ])
         if not n:
             return empty_div, False, None, html.Div()
-        filepath = os.path.normpath(filepath)
         if filepath:
+            filepath = os.path.normpath(filepath)
             if os.path.isdir(filepath):
                 try:
                     deployPage.fileExplorerInstance.root = filepath
-                    deployPage.terminal.writeln(f'$ Selected file {filepath}')
+                    if trigger_id == "file-explorer-button":
+                        deployPage.terminal.writeln(f'$ Selected file {filepath}')
                     alerts_list = []
                     if not gitUtils.git_is_installed():
                         deployPage.terminal.writeln(