+++ src/dashtools/dashboard/callbacks/router.py
@@ -5,10 +5,14 @@
 
 from dash import Dash, Input, Output, dcc
 
+from dashtools.dashboard.pages import explorePage
+
+from ..pages import explorePage
+
 try:
-    from dashtools.dashboard.pages import createPage, deployPage, errorPage, infoPage
+    from dashtools.dashboard.pages import createPage, deployPage, helpPage
 except ModuleNotFoundError:
-    from dashtools.dashboard.pages import createPage, deployPage, errorPage, infoPage
+    from ..pages import createPage, deployPage, helpPage
 
 
 def generate_callbacks(app: Dash):
@@ -16,14 +20,15 @@ def generate_callbacks(app: Dash):
     def render_page_content(pathname):
         if pathname == '/':
             # Setup url redirect to default page
-            return dcc.Location(id="url", pathname='/deploy')
+            return dcc.Location(id="url", pathname='/create')
         elif pathname == "/deploy":
-            deployPage.terminal.clear()
             return deployPage.render()
-        elif pathname == "/info":
-            return infoPage.render()
+        elif pathname == "/explore":
+            return explorePage.render()
         elif pathname == "/create":
-            createPage.terminal.clear()
             return createPage.render()
+        elif pathname == "/help":
+            return helpPage.render()
         else:
-            return errorPage.render()
+            # Can also implement custom error page here
+            return dcc.Location(id="url", pathname='/')