+++ src/dashtools/dashboard/components/sidebar.py
@@ -13,7 +13,7 @@ def render() -> html.Div:
                 DashIconify(icon='heroicons:command-line-20-solid',
                             height=60, style={'margin-bottom': '8px', 'margin-right': '5px'}),
                 html.H2("DashTools", className='dashtools-logo'),
-            ]),
+            ], style={"margin-top": "5px",}),
             dmc.Space(h=1, style={'margin-top': '-20px'}),
             html.H6(
                 "Application Management Dashboard",
@@ -26,7 +26,7 @@ def render() -> html.Div:
                     dbc.NavLink(
                         [
                             DashIconify(icon='akar-icons:plus',
-                                        style={'margin-right': '5px'}),
+                                        style={'margin-right': '5px', 'margin-bottom':'5px'}),
                             "Create"
                         ], href="/create", active="exact"),
                     dbc.NavLink(
@@ -35,57 +35,77 @@ def render() -> html.Div:
                                         style={'margin-right': '5px'}),
                             "Deploy"
                         ], href="/deploy", active="exact"),
-                    # dbc.NavLink(  # TODO
-                    #     [
-                    #         DashIconify(icon='akar-icons:info',
-                    #                     style={'margin-right': '5px'}),
-                    #         "Info"
-                    #     ], href="/info", active="exact"),
+                    dbc.NavLink(
+                        [
+                            DashIconify(icon='akar-icons:telescope',
+                                        style={'margin-right': '5px'}),
+                            "Explore (Preview)"
+                        ], href="/explore", active="exact", disabled=True),
+                    dbc.NavLink(
+                        [
+                            DashIconify(icon='akar-icons:question',
+                                        style={'margin-right': '5px'}),
+                            "Help"
+                        ], href="/help", active="exact"),
                 ],
                 vertical=True,
                 pills=True,
             ),
 
-            dmc.Space(style={'height': '460px'}),
-            html.Hr(),
             html.Div([
-                html.H6(
-                    [
-                        html.A(
-                            [
-                                DashIconify(
-                                    icon='logos:pypi',
-                                    width=20,
-                                    style={'margin-right': '10px'}),
-                                f'PyPi v{version.__version__}',
-                            ],
-                            href="https://pypi.org/project/dash-tools/", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
-                    ]),
-                html.H6(
-                    [
-                        html.A(
-                            [
-                                DashIconify(
-                                    icon='file-icons:readthedocs',
-                                    width=15,
-                                    style={'margin-right': '10px', 'margin-left': '5px'}),
-                                f'Read the Docs',
-                            ],
-                            href="https://dash-tools.readthedocs.io/en/latest/index.html", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
-                    ]),
-                html.H6(
-                    [
-                        html.A(
-                            [
-                                DashIconify(
-                                    icon='ant-design:github-filled',
-                                    width=20,
-                                    style={'margin-right': '8px', 'margin-left': '2px'}),
-                                f'GitHub',
-                            ],
-                            href="https://github.com/andrew-hossack/dash-tools", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
-                    ]),
-            ]),
+                html.Hr(style={'width':'200%'}),
+                html.Div([
+                    html.H6(
+                        [
+                            html.A(
+                                [
+                                    DashIconify(
+                                        icon='logos:pypi',
+                                        width=20,
+                                        style={'margin-right': '10px'}),
+                                    f'PyPi v{version.__version__}',
+                                ],
+                                href="https://pypi.org/project/dash-tools/", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
+                        ]),
+                    html.H6(
+                        [
+                            html.A(
+                                [
+                                    DashIconify(
+                                        icon='file-icons:readthedocs',
+                                        width=15,
+                                        style={'margin-right': '10px', 'margin-left': '5px'}),
+                                    f'Read the Docs',
+                                ],
+                                href="https://dash-tools.readthedocs.io/en/latest/index.html", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
+                        ]),
+                    html.H6(
+                        [
+                            html.A(
+                                [
+                                    DashIconify(
+                                        icon='ant-design:github-filled',
+                                        width=20,
+                                        style={'margin-right': '8px', 'margin-left': '2px'}),
+                                    f'GitHub',
+                                ],
+                                href="https://github.com/andrew-hossack/dash-tools", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
+                        ]),
+                    html.H6(
+                        [
+                            html.A(
+                                [
+                                    DashIconify(
+                                        icon='material-symbols:bug-report-outline-rounded',
+                                        width=20,
+                                        style={'margin-right': '8px', 'margin-left': '2px'}),
+                                    f'Report Bugs',
+                                ],
+                                href="https://github.com/andrew-hossack/dash-tools/issues/new/choose", target='_blank', style={'text-decoration': 'none', 'color': 'black', 'font-weight': 'lighter'}),
+                        ]),
+                ]),
+            ], style={'position':'absolute','bottom':'0', 'margin-bottom':'10px'})
+
         ],
         style={
             "position": "fixed",