+++ src/dashtools/templating/templates/default/preview.py
@@ -1,0 +1,26 @@
+# Returns a preview of the template, used for dashtools UI
+from dash import html, dcc
+import plotly.express as px
+import pandas as pd
+
+def render():
+    df = pd.DataFrame({
+        "Fruit": ["Apples", "Oranges", "Bananas", "Apples", "Oranges", "Bananas"],
+        "Amount": [4, 1, 2, 2, 4, 5],
+        "City": ["SF", "SF", "SF", "Montreal", "Montreal", "Montreal"]
+    })  
+
+    fig = px.bar(df, x="Fruit", y="Amount", color="City", barmode="group")
+
+    return html.Div(children=[
+            html.H1(children='Hello Dash'),
+
+            html.Div(children='''
+                Dash: A web application framework for your data.
+            '''),
+
+            dcc.Graph(
+                id='example-graph',
+                figure=fig
+            )
+        ])